+++ DesignToolbox/DesignToolbox (Light)/Views/Components/Actions/Button/ButtonsView.swift
@@ -19,6 +19,7 @@ struct ButtonsView: View {
 
     private static let kAllButtonAppaerances: [OUDSButton.Appearance] = [.default, .strong, .brand, .minimal, .negative]
     private static let kAllButtonStyles: [OUDSButton.Style] = [.default, .loading]
+    private static let kAllButtonSizes: [OUDSButton.Size] = [.default, .small]
 
     @Environment(\.theme) private var theme
 
@@ -30,37 +31,18 @@ struct ButtonsView: View {
     }
 
     @ViewBuilder
-    private func buttonSection(style: OUDSButton.Style, appearance: OUDSButton.Appearance) -> some View {
-        VStack(spacing: theme.spaces.scaledXsmallTablet) {
-            VStack(spacing: theme.spaces.scaled3xsmallTablet) {
-                Text(String(describing: style))
-                    .font(.caption)
-                    .fontWeight(.bold)
-                Text(String(describing: appearance))
-                    .font(.caption2)
-                    .foregroundColor(.secondary)
-            }
-            .multilineTextAlignment(.center)
-
-            VStack(spacing: theme.spaces.scaledXsmallTablet) {
-                OUDSButton(text: "Button", appearance: appearance, style: style) {}
-                OUDSButton(text: "Button", image: OUDSImage(asset: Image(systemName: "sun.min.fill")), appearance: appearance, style: style) {}
-                OUDSButton(image: OUDSImage(asset: Image(systemName: "sun.min.fill"), accessibilityLabel: "Button"), appearance: appearance, style: style) {}
-            }
-        }
-        .padding(theme.spaces.paddingInlineMedium)
-    }
-
-    @ViewBuilder
     private var watchLayout: some View {
         WatchVerticalLayout {
             ForEach(Self.kAllButtonStyles, id: \.self) { style in
                 Text("Style \(String(describing: style))").font(.headline)
-                ForEach(Self.kAllButtonAppaerances, id: \.self) { appearance in
-                    Text("Appearance \(String(describing: appearance))").font(.subheadline)
-                    OUDSButton(text: "Button", appearance: appearance, style: style) {}
-                    OUDSButton(text: "Button", image: OUDSImage(asset: Image(systemName: "sun.min.fill")), appearance: appearance, style: style) {}
-                    OUDSButton(image: OUDSImage(asset: Image(systemName: "sun.min.fill"), accessibilityLabel: "Button"), appearance: appearance, style: style) {}
+                ForEach(Self.kAllButtonSizes, id: \.self) { size in
+                    Text("Size \(String(describing: size))").font(.subheadline)
+                    ForEach(Self.kAllButtonAppaerances, id: \.self) { appearance in
+                        Text("Appearance \(String(describing: appearance))").font(.body)
+                        OUDSButton(text: "Button", appearance: appearance, style: style, size: size) {}
+                        OUDSButton(text: "Button", image: OUDSImage(asset: Image(systemName: "sun.min.fill")), appearance: appearance, style: style, size: size) {}
+                        OUDSButton(image: OUDSImage(asset: Image(systemName: "sun.min.fill"), accessibilityLabel: "Button"), appearance: appearance, style: style, size: size) {}
+                    }
                 }
             }
         }