+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- [Library] `button` component to version 3.3.0 (Orange-OpenSource/ouds-ios#1583)
 - [DesignToolbox] Swift Package `accessibility-statement-lib-ios` to v2.4.0-beta.1
 - [Library] **BREAKING**: `link` component to v2.4.0 (Orange-OpenSource/ouds-ios#1640)
 - [Library] `suggestion chip` and `filter chip` components to v1.5.0 (Orange-OpenSource/ouds-ios#1588)